Importing AFNI 0.20091020~dfsg.1 Imported from tarball: AFNI_2008_07_18_1710.tgz Original MD5: 7c47b01721f0484664577911298c67a4 AFNI claimed to be version: 0.20080718.1710

+++ mri_purger.c
@@ -3,36 +3,82 @@
 /*----------------------------------------------------------------------------*/
 static char *tmpdir = NULL ;
 
-/*! Function to get name of the directory to store TIM_* mri_purge() files. */
+static int test_tmpdir_write(void)
+{
+   char *un , *tf ; FILE *fp ; size_t bb=0 ;
+   if( tmpdir==NULL || *tmpdir=='\0' || !THD_is_directory(tmpdir) ) return 0 ;
+   un = UNIQ_idcode() ;
+   tf = malloc(strlen(un)+strlen(tmpdir)+8) ;
+   strcpy(tf,tmpdir) ; strcat(tf,"/") ; strcat(tf,un) ; free(un) ;
+   fp = fopen(tf,"w+b") ;
+   if( fp == NULL ){ free(tf) ; return 0 ; }
+   bb = fwrite( &bb , sizeof(size_t) , 1 , fp ) ;
+   fclose(fp) ; remove(tf) ; free(tf) ;
+   return ( bb == 1 ) ;
+}
+
+/*----------------------------------------------------------------------------*/
+/*! Function to get name of the directory to store TIM_* mri_purge() files.
+    The return pointer should NOT be free()-ed!
+*//*--------------------------------------------------------------------------*/
 
 char * mri_purge_get_tmpdir(void)
 {
    if( tmpdir == NULL ){
-                                             tmpdir = getenv( "TMPDIR" ) ;
-     if( tmpdir == NULL || *tmpdir == '\0' ) tmpdir = getenv( "TEMPDIR" ) ;
-     if( tmpdir == NULL || *tmpdir == '\0' ) tmpdir = "/tmp" ;
-     if( !THD_is_directory(tmpdir) )         tmpdir = "." ;
+                                tmpdir = getenv( "TMPDIR" ) ;
+     if( !test_tmpdir_write() ) tmpdir = getenv( "TEMPDIR" ) ;
+     if( !test_tmpdir_write() ) tmpdir = "/tmp" ;
+     if( !test_tmpdir_write() ) tmpdir = "." ;    /* the last resort */
    }
    return tmpdir ;
 }
+
+/*----------------------------------------------------------------------------*/
+/*! Function to return a unique name for temporary file, with the
+    given prefix attached (if it isn't NULL, that is).
+    The return pointer should be free()-ed when you are done with it.
+*//*--------------------------------------------------------------------------*/
+
+char * mri_get_tempfilename( char *pref )
+{
+   char *tdir , *unam , *fnam ; int nn ;
+
+   tdir = mri_purge_get_tmpdir() ;
+
+   do{
+     unam = UNIQ_idcode() ;                        /* unique part of filename */
+     nn   = (pref != NULL) ? strlen(pref) : 0 ;           /* length of prefix */
+     fnam = (char *)malloc(sizeof(char)*(strlen(tdir)+strlen(unam)+nn+4)) ;
+     strcpy(fnam,tdir) ;                    /* start with temp directory name */
+     nn = strlen(fnam) ; if( fnam[nn-1] != '/' ) strcat(fnam,"/") ;
+     if( pref != NULL ){ strcat(fnam,pref); strcat(fnam,"_"); } /* add prefix */
+     strcat(fnam,unam) ; free(unam) ;                 /* append unique string */
+     nn = THD_is_ondisk(fnam) ;                        /* should never happen */
+     if( nn ) free(fnam) ;
+   } while(nn) ;                                         /* should never loop */
+
+   return fnam ;
+}
+
 
 /*----------------------------------------------------------------------------*/
 static char tsuf[8] = "\0" ;
 
 /*! Function to set TIM suffix for this process */
 
-static void purge_set_tsuf(void)  /* 01 Feb 2007 */
+static void purge_set_tsuf(int v)  /* 01 Feb 2007 */
 {
    char *un ;
    if( tsuf[0] != '\0' ) return ;
    un = UNIQ_idcode();
    tsuf[0] = un[5] ; tsuf[1] = un[6] ; tsuf[2] = un[7] ; tsuf[3] = '\0' ;
-   INFO_message("temp files: if program crashes, do /bin/rm -f %s/TIM_%s*",
-                mri_purge_get_tmpdir() , tsuf ) ;
+   if( v )
+     INFO_message("temp files: if program crashes, do /bin/rm -f %s/TIM_%s*",
+                  mri_purge_get_tmpdir() , tsuf ) ;
    return ;
 }
 
-char * mri_purge_get_tsuf(void){ purge_set_tsuf(); return tsuf; }
+char * mri_purge_get_tsuf(void){ purge_set_tsuf(0); return tsuf; }
 
 /*----------------------------------------------------------------------------*/
 /* Functions to save a list of TIM_* mri_purge() files, so that they can
@@ -119,7 +165,7 @@ ENTRY("mri_purge") ;
 
    /* make up a unique name for the purge file */
 
-   pg        = mri_purge_get_tmpdir() ; purge_set_tsuf() ;
+   pg        = mri_purge_get_tmpdir() ; purge_set_tsuf(1) ;
    im->fname = malloc(strlen(pg)+64) ;
    un = UNIQ_idcode();
    un[0] = 'T'     ; un[1] = 'I'     ; un[2] = 'M'     ; un[3] = '_' ;
@@ -167,7 +213,7 @@ void mri_unpurge( MRI_IMAGE *im )
 ENTRY("mri_unpurge") ;
 
    if( im == NULL || !MRI_IS_PURGED(im) ) EXRETURN ;
-   iar = mri_data_pointer_unvarnished(im) ; if( iar != NULL ) EXRETURN ;
+   iar = im->im ; if( iar != NULL ) EXRETURN ;
 
    fp = fopen( im->fname , "rb" ) ;
    if( fp == NULL ){
